llvm: fix rpath bug when compiling on Cray
Compiling LLVM with GCC fails on a Cray with an error:

No valid ELF RPATH or RUNPATH entry exists in the file

This occurs because the CMake flag '-DCMAKE_BUILD_WITH_INSTALL_RPATH=1', which
fixes this error, is set only if llvm is being compiled on a 'linux' platform.
The flag is not set if compiling on a 'cray' platform. So adjust the logic to
add 'cray' to the list of platforms that should use this CMake flag.