Add Cray for compiler flags
In the case of lua - static_to_shared method would not work on Cray
because the compiler flags would not be used to create the shared object
since it only checks for linux and darwin. Since Cray is pretty much a
linux box with Cray modifications, we can use those same flags.